# src/main/resources/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Ousmane Vincent Dione",
    "description": "OpenApi documentation for Stock Management",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ousmane Vincent",
      "url": "https://ndiolfouta.com",
      "email": "dioneousmanevincent@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local ENV"
    },
    {
      "url": "https://ndiolfouta.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/fournisseurs": {
      "post": {
        "tags": [
          "fournisseur-api"
        ],
        "operationId": "save_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FournisseurDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entreprises": {
      "post": {
        "tags": [
          "entreprise-api"
        ],
        "operationId": "save_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntrepriseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntrepriseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeFournisseurs": {
      "post": {
        "tags": [
          "commande-fournisseur-api"
        ],
        "operationId": "save_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandeFournisseurDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients": {
      "post": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "save_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandeClientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commandeClients/updateQuantite/{idCommande}/{idLigneCommande}/{quantite}": {
      "patch": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "updateQuantiteCommande",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "quantite",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients/updateEtat/{idCommande}/{etatCommande}": {
      "patch": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "updateEtatCommande",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "etatCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "EN_PREPARATION",
                "VALIDEE",
                "LIVREE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients/updateClient/{idCommande}/{idClient}": {
      "patch": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "updateClient",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients/updateArticle/{idCommande}/{idLigneCommande}/{idArticle}": {
      "patch": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fournisseurs/{idFournisseur}": {
      "get": {
        "tags": [
          "fournisseur-api"
        ],
        "operationId": "getFournisseurById",
        "parameters": [
          {
            "name": "idFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/entreprises/{idEntreprise}": {
      "get": {
        "tags": [
          "entreprise-api"
        ],
        "operationId": "getEntrepriseById",
        "parameters": [
          {
            "name": "idEntreprise",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntrepriseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commandeFournisseurs/{idCFournisseur}": {
      "get": {
        "tags": [
          "commande-fournisseur-api"
        ],
        "operationId": "getCommandeFournisseurById",
        "parameters": [
          {
            "name": "idCFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commandeClients/{idCommande}": {
      "get": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "getCommandeClientById",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients/ligneCommande/{idCommande}": {
      "get": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "findAllLignesCommandesClientByCommandeClientId",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneCommandeClientDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fournisseurs/delete/{idFournisseur}": {
      "delete": {
        "tags": [
          "fournisseur-api"
        ],
        "operationId": "deleteFournisseur",
        "parameters": [
          {
            "name": "idFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entreprises/delete/{idEntreprise}": {
      "delete": {
        "tags": [
          "entreprise-api"
        ],
        "operationId": "deleteEntreprise",
        "parameters": [
          {
            "name": "idEntreprise",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeFournisseurs/delete/{idCFournisseur}": {
      "delete": {
        "tags": [
          "commande-fournisseur-api"
        ],
        "operationId": "deleteCommandeFournisseur",
        "parameters": [
          {
            "name": "idCFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients/deleteArticle/{idCommande}/{idLigneCommande}": {
      "delete": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commandeClients/delete/{idCommande}": {
      "delete": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "deleteCommandeClient",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdresseDto": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostal": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
      "EntrepriseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "codeFiscal": {
            "type": "string"
          },
          "siteWeb": {
            "type": "string"
          },
          "adresseDto": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "utilisateurDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UtilisateurDto"
            }
          }
        }
      },
      "UtilisateurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "entrepriseDto": {
            "$ref": "#/components/schemas/EntrepriseDto"
          }
        }
      },
      "CommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "fournisseurDto": {
            "$ref": "#/components/schemas/FournisseurDto"
          }
        }
      },
      "FournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "adresseDto": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "commandeFournisseurDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandeFournisseurDto"
            }
          }
        }
      },
      "Adresse": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostal": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/Adresse"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          }
        }
      },
      "CommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "clientDto": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "etat": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          }
        }
      },
...
      "ArticleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codeArticle": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "prixUnitaireHt": {
            "type": "number"
          },
          "tauxTval": {
            "type": "number"
          },
          "prixUnitaireTtc": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "categoryDto": {
            "$ref": "#/components/schemas/CategoryDto"
          }
        }
      },
      "LigneCommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleDto": {
            "$ref": "#/components/schemas/ArticleDto"
          },
          "commandeClientDto": {
            "$ref": "#/components/schemas/CommandeClientDto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Gestion stock auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}